Fix implicit casting

+++ dfa.cpp
@@ -45,8 +45,8 @@ int main() {
 
       xt::xarray<float> y = network.forward(x);
 
-      loss += error.with(t).forward(y) * batchsize;
-      acc += F::evaluation::accuracy(t, y) * batchsize;
+      loss += error.with(t).forward(y) * static_cast<float>(batchsize);
+      acc += F::evaluation::accuracy(t, y) * static_cast<float>(batchsize);
 
       network.backward(error.grads());
       network.update();
@@ -56,8 +56,10 @@ int main() {
     dataset.for_each(batchsize, train);
 
     std::cout << "\rEpoch " << std::right << std::setfill(' ') << std::setw(2)
-              << epoch + 1 << " Loss: " << std::scientific << loss / n_train
-              << " Accuracy: " << acc / n_train << std::endl;
+              << epoch + 1 << " Loss: " << std::scientific
+              << loss / static_cast<float>(n_train)
+              << " Accuracy: " << acc / static_cast<float>(n_train)
+              << std::endl;
   }
 
   return 0;
